+++ src/Components/TaskCard.jsx
@@ -3,7 +3,7 @@ import { React } from "react";
 
 export default function ({ task }) {
   return (
-    <Link to={"/doneTask/" + task.id}>
+    <Link to={"/tasks/" + task.id}>
       <div className="my-5 rounded-3xl bg-gradient-to-br from-bright_accent to-accent p-6">
         <div className="flex justify-center text-lg font-bold">{task.name}</div>
