+++ src/main.jsx
@@ -7,9 +7,9 @@ import Landing from "./Pages/Landing";
 import NavbarRouteContainer from "./Pages/NavbarRouteContainer";
 import Home from "./Pages/Home";
 import Tasks from "./Pages/Tasks";
-import CreateTask from "./Pages/CreateTask";
-import DoneTask from "./Pages/DoneTask";
-import Timer from "./Pages/Timer";
+import TaskCreate from "./Pages/TaskCreate";
+import Task from "./Pages/Task";
+import TaskRun from "./Pages/TaskRun";
 import Statistics from "./Pages/Statistics";
 import AllTrainings from "./Pages/AllTrainings";
 import "./index.css";
@@ -37,16 +37,16 @@ const router = createBrowserRouter(
               element: <Tasks />,
             },
             {
-              path: "/createTask",
-              element: <CreateTask />,
+              path: "/task_create",
+              element: <TaskCreate />,
             },
             {
-              path: "/doneTask/:id",
-              element: <DoneTask />,
+              path: "/tasks/:id",
+              element: <Task />,
             },
             {
-              path: "/timer",
-              element: <Timer />,
+              path: "/task_run",
+              element: <TaskRun />,
             },
             {
               path: "/statistics",
